add delete action, add task controller & templates, update entity
