+++ templates/list/edit.html.twig
@@ -1,0 +1,10 @@
+{% extends 'base.html.twig' %}
+{% block title %}
+    Edit your Todos
+{% endblock %}
+{% block body %}
+    <h2 class=page-header>Edit Todos</h2>
+    {{form_start(form)}}
+    {{form_widget(form)}}
+    {{form_end(form)}}
+{% endblock %}
